+++ src/Defer/DeferrableDirective.php
@@ -9,16 +9,27 @@ use GraphQL\Type\Definition\ResolveInfo;
 use GraphQL\Language\AST\NonNullTypeNode;
 use Nuwave\Lighthouse\Schema\AST\ASTHelper;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
-use Nuwave\Lighthouse\Support\Contracts\Directive;
 use Nuwave\Lighthouse\Schema\Directives\BaseDirective;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
+use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 
-class DeferrableDirective extends BaseDirective implements Directive, FieldMiddleware
+class DeferrableDirective extends BaseDirective implements DefinedDirective, FieldMiddleware
 {
     const NAME = 'deferrable';
     const THE_DEFER_DIRECTIVE_CANNOT_BE_USED_ON_A_ROOT_MUTATION_FIELD = 'The @defer directive cannot be used on a root mutation field.';
     const THE_DEFER_DIRECTIVE_CANNOT_BE_USED_ON_A_NON_NULLABLE_FIELD = 'The @defer directive cannot be used on a Non-Nullable field.';
+
+    public static function definition(): string
+    {
+        return /* @lang GraphQL */ <<<'SDL'
+"""
+Do not use this directive directly, it is automatically added to the schema
+when using the defer extension.
+"""
+directive @deferrable on FIELD_DEFINITION
+SDL;
+    }
 
     /**
      * @var \Nuwave\Lighthouse\Defer\Defer