+++ src/Schema/Directives/AllDirective.php
@@ -7,8 +7,9 @@ use Illuminate\Database\Eloquent\Collection;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
+use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 
-class AllDirective extends BaseDirective implements FieldResolver
+class AllDirective extends BaseDirective implements DefinedDirective, FieldResolver
 {
     /**
      * Name of the directive.
@@ -18,6 +19,24 @@ class AllDirective extends BaseDirective implements FieldResolver
     public function name(): string
     {
         return 'all';
+    }
+
+    public static function definition(): string
+    {
+        return /* @lang GraphQL */ <<<'SDL'
+directive @all(
+  """
+  Specify the class name of the model to use.
+  This is only needed when the default model resolution does not work.
+  """
+  model: String
+
+  """
+  Apply scopes to the underlying query.
+  """
+  scopes: [String!]
+) on FIELD_DEFINITION
+SDL;
     }
 
     /**