+++ src/Schema/Directives/AuthDirective.php
@@ -6,8 +6,9 @@ use Illuminate\Contracts\Auth\Authenticatable;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Illuminate\Contracts\Auth\Factory as AuthFactory;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
+use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 
-class AuthDirective extends BaseDirective implements FieldResolver
+class AuthDirective extends BaseDirective implements DefinedDirective, FieldResolver
 {
     /**
      * @var \Illuminate\Contracts\Auth\Factory
@@ -33,6 +34,21 @@ class AuthDirective extends BaseDirective implements FieldResolver
     public function name(): string
     {
         return 'auth';
+    }
+
+    public static function definition(): string
+    {
+        return /* @lang GraphQL */ <<<'SDL'
+"""
+Return the currently authenticated user as the result of a query.
+"""
+directive @auth(
+  """
+  Use a particular guard to retreive the user.
+  """
+  guard: String
+) on FIELD_DEFINITION
+SDL;
     }
 
     /**