+++ src/Schema/Directives/OrderByDirective.php
@@ -9,10 +9,11 @@ use GraphQL\Language\AST\InputValueDefinitionNode;
 use GraphQL\Language\AST\ObjectTypeDefinitionNode;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Support\Contracts\ArgManipulator;
+use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 use Nuwave\Lighthouse\Support\Contracts\ArgBuilderDirective;
 use Nuwave\Lighthouse\Support\Contracts\ArgDirectiveForArray;
 
-class OrderByDirective implements ArgBuilderDirective, ArgDirectiveForArray, ArgManipulator
+class OrderByDirective implements ArgBuilderDirective, ArgDirectiveForArray, ArgManipulator, DefinedDirective
 {
     /**
      * Name of the directive.
@@ -22,6 +23,16 @@ class OrderByDirective implements ArgBuilderDirective, ArgDirectiveForArray, Arg
     public function name(): string
     {
         return 'orderBy';
+    }
+
+    public static function definition(): string
+    {
+        return /* @lang GraphQL */ <<<'SDL'
+"""
+Sort a result list by one or more given fields.
+"""
+directive @orderBy on ARGUMENT_DEFINITION
+SDL;
     }
 
     /**